+++ src/sprite/mod.rs
@@ -1,4 +1,4 @@
-use std::collections::BTreeMap;
+use std::collections::{BTreeMap, HashMap};
 use std::fs::File;
 use std::io::Write;
 use std::path::Path;
@@ -7,10 +7,141 @@ use crunch::{Item, PackedItem, PackedItems, Rotation};
 use multimap::MultiMap;
 use oxipng::optimize_from_memory;
 use resvg::tiny_skia::{Pixmap, PixmapPaint, Transform};
-use resvg::usvg::Tree;
+use resvg::usvg::{NodeExt, Rect, Tree};
 use serde::Serialize;
 
+use self::serialize::{serialize_rect, serialize_stretch_x_area, serialize_stretch_y_area};
 use crate::error::Error;
+
+mod serialize;
+
+/// A single icon within a spritesheet.
+///
+/// A sprite is a rectangular icon stored as an SVG image and converted to a bitmap. The bitmap is
+/// saved to a spritesheet.
+#[derive(Clone)]
+pub struct Sprite {
+    /// Parsed source SVG image.
+    pub tree: Tree,
+    /// Ratio determining the size the destination pixels compared to the source pixels. A ratio of
+    /// 2 means the bitmap will be scaled to be twice the size of the SVG image.
+    pub pixel_ratio: u8,
+}
+
+impl Sprite {
+    /// Generate a bitmap image from the sprite's SVG tree.
+    ///
+    /// The bitmap is generated at the sprite's [pixel ratio](Self::pixel_ratio).
+    pub fn pixmap(&self) -> Option<Pixmap> {
+        let rtree = resvg::Tree::from_usvg(&self.tree);
+        let pixmap_size = rtree.size.to_int_size().scale_by(self.pixel_ratio as f32)?;
+        let mut pixmap = Pixmap::new(pixmap_size.width(), pixmap_size.height())?;
+        let render_ts = Transform::from_scale(self.pixel_ratio.into(), self.pixel_ratio.into());
+        rtree.render(render_ts, &mut pixmap.as_mut());
+        Some(pixmap)
+    }
+
+    /// Metadata for a [stretchable icon].
+    ///
+    /// Describes the content area of an icon as a [`Rect`]. The metadata comes from the bounding
+    /// box of an element in the SVG image that has the id `mapbox-content`.
+    ///
+    /// Most icons do not specify a content area. But if it is present and the MapLibre/Mapbox map
+    /// symbol uses [`icon-text-fit`], the symbol's text will be fitted inside this content box.
+    ///
+    /// [stretchable icon]: https://github.com/mapbox/mapbox-gl-js/issues/8917
+    /// [`icon-text-fit`]: https://maplibre.org/maplibre-style-spec/layers/#layout-symbol-icon-text-fit
+    pub fn content_area(&self) -> Option<Rect> {
+        self.get_node_bbox("mapbox-content")
+    }
+
+    /// Metadata for a [stretchable icon].
+    ///
+    /// Describes the horizontal position of areas that can be stretched. There may be multiple
+    /// areas. The metadata comes from the bounding boxes of elements in the SVG image that have
+    /// ids like `mapbox-stretch-x-1`. Although the entire bounding box is provided, only the left
+    /// and right edges are stored in the index file and used by MapLibre/Mapbox to define the
+    /// stretchable area.
+    ///
+    /// Most icons do not specify stretchable areas. See also [`Sprite::content_area`].
+    ///
+    /// [stretchable icon]: https://github.com/mapbox/mapbox-gl-js/issues/8917
+    pub fn stretch_x_areas(&self) -> Option<Vec<Rect>> {
+        let mut values = vec![];
+        // First look for an SVG element with the id `mapbox-stretch-x`.
+        if let Some(rect) = self.get_node_bbox("mapbox-stretch-x") {
+            values.push(rect);
+        }
+        // Next look for SVG elements with ids like `mapbox-stretch-x-1`. As soon as one is missing,
+        // stop looking.
+        for i in 1.. {
+            if let Some(rect) = self.get_node_bbox(format!("mapbox-stretch-x-{}", i).as_str()) {
+                values.push(rect);
+            } else {
+                break;
+            }
+        }
+        if values.is_empty() {
+            // If there are no SVG elements with `mapbox-stretch-x` ids, check for an element with
+            // the id `mapbox-stretch`. That's a shorthand for stretch-x and stretch-y. If that
+            // exists, use its horizontal coordinates.
+            self.get_node_bbox("mapbox-stretch").map(|rect| vec![rect])
+        } else {
+            Some(values)
+        }
+    }
+
+    /// Metadata for a [stretchable icon].
+    ///
+    /// Describes the vertical position of areas that can be stretched. There may be multiple areas.
+    /// The metadata comes from the bounding boxes of elements in the SVG image that have ids like
+    /// `mapbox-stretch-y-1`. Although the entire bounding box is provided, only the top and bottom
+    /// edges are stored in the index file and used by MapLibre/Mapbox to define the stretchable
+    /// area.
+    ///
+    /// Most icons do not specify stretchable areas. See also [`Sprite::content_area`].
+    ///
+    /// [stretchable icon]: https://github.com/mapbox/mapbox-gl-js/issues/8917
+    pub fn stretch_y_areas(&self) -> Option<Vec<Rect>> {
+        let mut values = vec![];
+        // First look for an SVG element with the id `mapbox-stretch-y`.
+        if let Some(rect) = self.get_node_bbox("mapbox-stretch-y") {
+            values.push(rect);
+        }
+        // Next look for SVG elements with ids like `mapbox-stretch-y-1`. As soon as one is missing,
+        // stop looking.
+        for i in 1.. {
+            if let Some(rect) = self.get_node_bbox(format!("mapbox-stretch-y-{}", i).as_str()) {
+                values.push(rect);
+            } else {
+                break;
+            }
+        }
+        if values.is_empty() {
+            // If there are no SVG elements with `mapbox-stretch-x` ids, check for an element with
+            // the id `mapbox-stretch`. That's a shorthand for stretch-x and stretch-y. If that
+            // exists, use its vertical coordinates.
+            self.get_node_bbox("mapbox-stretch").map(|rect| vec![rect])
+        } else {
+            Some(values)
+        }
+    }
+
+    /// Find a node in the SVG tree with a given id, and return its bounding box with coordinates
+    /// multiplied by the sprite's pixel ratio.
+    fn get_node_bbox(&self, id: &str) -> Option<Rect> {
+        self.tree.node_by_id(id)?.calculate_bbox().map(|bbox| {
+            let ratio = self.pixel_ratio as f32;
+            Rect::from_ltrb(
+                bbox.left() * ratio,
+                bbox.top() * ratio,
+                bbox.right() * ratio,
+                bbox.bottom() * ratio,
+            )
+            .unwrap()
+        })
+    }
+}
 
 /// A description of a sprite image within a spritesheet. Used for the JSON output required by a
 /// Mapbox Style Specification [index file].
@@ -24,13 +155,28 @@ pub struct SpriteDescription {
     pub width: u32,
     pub x: u32,
     pub y: u32,
+    #[serde(
+        skip_serializing_if = "Option::is_none",
+        serialize_with = "serialize_rect"
+    )]
+    pub content: Option<Rect>,
+    #[serde(
+        skip_serializing_if = "Option::is_none",
+        serialize_with = "serialize_stretch_x_area"
+    )]
+    pub stretch_x: Option<Vec<Rect>>,
+    #[serde(
+        skip_serializing_if = "Option::is_none",
+        serialize_with = "serialize_stretch_y_area"
+    )]
+    pub stretch_y: Option<Vec<Rect>>,
 }
 
 /// Builder pattern for `Spritesheet`: construct a `Spritesheet` object using calls to a builder
 /// helper.
 #[derive(Default)]
 pub struct SpritesheetBuilder {
-    sprites: Option<BTreeMap<String, Pixmap>>,
+    sprites: Option<BTreeMap<String, Sprite>>,
     references: Option<MultiMap<String, String>>,
     pixel_ratio: u8,
 }
@@ -44,7 +190,7 @@ impl SpritesheetBuilder {
         }
     }
 
-    pub fn sprites(&mut self, sprites: BTreeMap<String, Pixmap>) -> &mut Self {
+    pub fn sprites(&mut self, sprites: BTreeMap<String, Sprite>) -> &mut Self {
         self.sprites = Some(sprites);
         self
     }
@@ -63,7 +209,7 @@ impl SpritesheetBuilder {
                 let mut references = MultiMap::new();
                 let mut names_for_sprites: BTreeMap<Vec<u8>, String> = BTreeMap::new();
                 for (name, sprite) in sprites {
-                    let sprite_data = sprite.encode_png().unwrap();
+                    let sprite_data = sprite.pixmap().unwrap().encode_png().unwrap();
                     if let Some(existing_sprite_name) = names_for_sprites.get(&sprite_data) {
                         references.insert(existing_sprite_name.clone(), name.clone());
                     } else {
@@ -98,15 +244,20 @@ pub struct Spritesheet {
 
 impl Spritesheet {
     pub fn new(
-        sprites: BTreeMap<String, Pixmap>,
+        sprites: BTreeMap<String, Sprite>,
         references: MultiMap<String, String>,
         pixel_ratio: u8,
     ) -> Option<Self> {
+        let pixmaps: HashMap<&String, Pixmap> = sprites
+            .iter()
+            .map(|(name, sprite)| (name, sprite.pixmap().unwrap()))
+            .collect();
         // The items are the rectangles that we want to pack into the smallest space possible. We
         // don't need to pass the pixels themselves, just the unique name for each sprite.
         let items: Vec<Item<String>> = sprites
-            .iter()
-            .map(|(name, image)| {
+            .keys()
+            .map(|name| {
+                let image = pixmaps.get(name).unwrap();
                 Item::new(
                     name.clone(),
                     image.width() as usize,
@@ -115,10 +266,13 @@ impl Spritesheet {
                 )
             })
             .collect();
-        // Minimum area required for the spreadsheet (i.e. 100% coverage).
+        // Minimum area required for the spritesheet (i.e. 100% coverage).
         let min_area: usize = sprites
-            .values()
-            .map(|i| i.width() as usize * i.height() as usize)
+            .keys()
+            .map(|name| {
+                let image = pixmaps.get(name).unwrap();
+                image.width() as usize * image.height() as usize
+            })
             .sum();
         match crunch::pack_into_po2(min_area * 10, items) {
             Ok(PackedItems { items: packed, .. }) => {
@@ -150,7 +304,7 @@ impl Spritesheet {
                     spritesheet.draw_pixmap(
                         rect.x as i32,
                         rect.y as i32,
-                        sprite.as_ref(),
+                        pixmaps.get(sprite_name).unwrap().as_ref(),
                         &pixmap_paint,
                         pixmap_transform,
                         None,
@@ -163,6 +317,9 @@ impl Spritesheet {
                             pixel_ratio,
                             x: rect.x as u32,
                             y: rect.y as u32,
+                            content: sprite.content_area(),
+                            stretch_x: sprite.stretch_x_areas(),
+                            stretch_y: sprite.stretch_y_areas(),
                         },
                     );
                     // If multiple names are used for a unique sprite, insert an entry in the index
@@ -179,6 +336,9 @@ impl Spritesheet {
                                     pixel_ratio,
                                     x: rect.x as u32,
                                     y: rect.y as u32,
+                                    content: sprite.content_area(),
+                                    stretch_x: sprite.stretch_x_areas(),
+                                    stretch_y: sprite.stretch_y_areas(),
                                 },
                             );
                         }
@@ -282,15 +442,11 @@ pub fn sprite_name<P1: AsRef<Path>, P2: AsRef<Path>>(path: P1, base_path: P2) ->
 ///
 /// The bitmap is generated at the given pixel ratio. A ratio of 2 means the bitmap image will be
 /// scaled to be double the size of the SVG image.
+#[deprecated(since = "0.9.0", note = "use `Sprite::pixmap()`instead")]
 pub fn generate_pixmap_from_svg(svg: &Tree, pixel_ratio: u8) -> Option<Pixmap> {
-    let rtree = resvg::Tree::from_usvg(svg);
-    let pixmap_size = rtree
-        .size
-        .to_int_size()
-        .scale_by(pixel_ratio as f32)
-        .unwrap();
-    let mut pixmap = Pixmap::new(pixmap_size.width(), pixmap_size.height()).unwrap();
-    let render_ts = Transform::from_scale(pixel_ratio as f32, pixel_ratio as f32);
-    rtree.render(render_ts, &mut pixmap.as_mut());
-    Some(pixmap)
+    Sprite {
+        tree: svg.clone(),
+        pixel_ratio,
+    }
+    .pixmap()
 }